Fix memory leaks.
