+++ src/BUILD
@@ -49,5 +49,6 @@ cc_binary(
     deps = [
         ":thumbnailer_cc_proto",
         ":thumbnailer_lib",
+        "//src/utils:thumbnailer_utils",
     ],
 )
