+++ src/utils/thumbnailer_utils.cc
@@ -16,6 +16,16 @@
 
 namespace libwebp {
 
+void WebPPictureDelete(WebPPicture* picture) {
+  WebPPictureFree(picture);
+  delete picture;
+}
+
+void WebPDataDelete(WebPData* webp_data) {
+  WebPDataClear(webp_data);
+  delete webp_data;
+}
+
 UtilsStatus AnimData2Frames(WebPData* const webp_data,
                             std::vector<Frame>* const frames) {
   std::unique_ptr<WebPAnimDecoder, void (*)(WebPAnimDecoder*)> dec(
@@ -75,7 +85,7 @@ UtilsStatus AnimData2PSNR(const std::vector<Frame>& original_frames,
     return kGenericError;
   }
 
-  int new_frame_index = 0;
+  std::vector<Frame>::size_type new_frame_index = 0;
   for (const Frame& original_frame : original_frames) {
     // Check if the next frame of new_frames matches original_frame,
     // based on their timestamps.
@@ -121,7 +131,6 @@ UtilsStatus CompareThumbnail(const std::vector<Frame>& original_frames,
   if (diff == nullptr) return kMemoryError;
 
   ThumbnailStatsPSNR stats_1, stats_2;
-  UtilsStatus status;
   CHECK_UTILS_STATUS(AnimData2PSNR(original_frames, webp_data_1, &stats_1));
   CHECK_UTILS_STATUS(AnimData2PSNR(original_frames, webp_data_2, &stats_2));
 
@@ -159,7 +168,7 @@ void PrintThumbnailStatsPSNR(const ThumbnailStatsPSNR& stats,
   std::cout << std::fixed << std::showpoint;
   std::cout << std::setprecision(3);
 
-  for (int i = 0; i < stats.psnr.size(); ++i) {
+  for (std::vector<float>::size_type i = 0; i < stats.psnr.size(); ++i) {
     std::cout << stats.psnr[i] << ' ';
   }
   std::cout << std::endl;
@@ -194,7 +203,7 @@ void PrintThumbnailDiffPSNR(const ThumbnailDiffPSNR& diff,
   std::cout << std::fixed << std::showpoint;
   std::cout << std::setprecision(3);
 
-  for (int i = 0; i < diff.psnr_diff.size(); ++i) {
+  for (std::vector<float>::size_type i = 0; i < diff.psnr_diff.size(); ++i) {
     std::cout << diff.psnr_diff[i] << ' ';
   }
   std::cout << std::endl;
